+++ lib/models/member_model.dart
