+++ lib/provider/abacus_registration_provider.dart
@@ -3,9 +3,8 @@ import 'dart:io';
 
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
-import 'package:http/http.dart' as http;
 
-import '../models/memberModel.dart';
+import '../models/member_model.dart';
 import '../utils/network_engine.dart';
 
 class AbacusRegistrationProvider with ChangeNotifier {
@@ -17,7 +16,6 @@ class AbacusRegistrationProvider with ChangeNotifier {
     required String teamLeaderScholarID,
     required List<MemberModel> members,
   }) async {
-    final dio = await NetworkEngine.getDio();
     List memberScholarIDs = members.map((e) => e.scholarId).toList();
     Map parameters = {
       "teamName": teamName,
@@ -28,9 +26,9 @@ class AbacusRegistrationProvider with ChangeNotifier {
     var jsonData = await jsonEncode(parameters);
 
     print(jsonData);
-    print('${apiUrl}${eventId}');
+    print('$apiUrl$eventId');
     var res = await (await NetworkEngine.getDio()).post(
-      '${apiUrl}${eventId}',
+      '$apiUrl$eventId',
       options: Options(
         followRedirects: false,
         validateStatus: (status) => true,
