+++ lib/state_handlers/theme/brightness/dark.dart
@@ -5,9 +5,9 @@ ThemeData darkTheme = ThemeData(
   primaryColor: Pallet.darkPrimaryColor,
   scaffoldBackgroundColor: Pallet.darkPrimaryColor,
   fontFamily: 'Archivo',
-  backgroundColor: Pallet.darkCardBGColor,
   canvasColor: Pallet.darkPrimaryColor,
   colorScheme: ColorScheme.fromSwatch(
+    backgroundColor: Pallet.darkCardBGColor,
     primarySwatch: Pallet.primarySwatch,
     brightness: Brightness.dark,
   ).copyWith(secondary: Pallet.darkSecondaryColor),
