+++ lib/state_handlers/theme/brightness/light.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:cssapp/configs/configurations/pallet.dart';
 
 ThemeData lightTheme = ThemeData(
-  backgroundColor: Pallet.lightSecondaryColor,
   scaffoldBackgroundColor: Pallet.lightPrimaryColor,
   fontFamily: 'Archivo',
   canvasColor: Pallet.lightPrimaryColor,
@@ -11,6 +10,7 @@ ThemeData lightTheme = ThemeData(
     displayColor: Pallet.lightSecondaryColor,
   ),
   colorScheme: ColorScheme.fromSwatch(
+    backgroundColor: Pallet.lightSecondaryColor,
     primarySwatch: Pallet.primarySwatch,
     brightness: Brightness.light,
   ).copyWith(secondary: Pallet.lightSecondaryColor),
