+++ lib/utils/storage_handler.dart
@@ -1,7 +1,6 @@
 // ignore_for_file: avoid_print
 
 import 'dart:ui';
-import 'package:flutter/scheduler.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class StorageHandler {
@@ -10,13 +9,12 @@ class StorageHandler {
   // ----------------------------- Getter Methods -----------------------------
   bool isDarkTheme() {
     return _preferences?.getBool('isDarkTheme') ??
-        SchedulerBinding.instance.window.platformBrightness == Brightness.dark;
+        PlatformDispatcher.instance.platformBrightness == Brightness.dark;
   }
 
   // ----------------------------- Setter Methods -----------------------------
   Future<void> toggleDarkTheme() async {
     await _preferences?.setBool('isDarkTheme', !isDarkTheme());
-    // print(isDarkTheme());
   }
 
   Future<void> initPreferences() async {
