+++ lib/widgets/member_card.dart
@@ -2,6 +2,7 @@
 
 import 'package:cssapp/configs/configs.dart';
 import 'package:cssapp/widgets/member_connect_slide_up.dart';
+import 'package:cssapp/widgets/member_name_card.dart';
 import 'package:fade_shimmer/fade_shimmer.dart';
 
 import 'social_icons.dart';
@@ -59,282 +60,9 @@ class MemberCard extends StatelessWidget {
             ),
           );
         },
-        child: NameCard(imageLink: imageLink, role: role, name: name),
+        child: MemberNameCard(imageLink: imageLink, role: role, name: name),
       ),
     );
   }
 }
 
-class NameCard extends StatelessWidget {
-  const NameCard({
-    Key? key,
-    required this.imageLink,
-    required this.role,
-    required this.name,
-  }) : super(key: key);
-
-  final String imageLink;
-  final String role;
-  final String name;
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 230,
-      height: 250,
-      child: Stack(
-        children: [
-          Positioned(
-            top: 30,
-            left: 5,
-            child: Container(
-              decoration: BoxDecoration(
-                color: Theme.of(context).backgroundColor,
-                borderRadius: BorderRadius.circular(10),
-              ),
-              width: 220,
-              height: 270,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.end,
-                children: [
-                  imageLink.isEmpty
-                      ? const Padding(
-                          padding: EdgeInsets.symmetric(horizontal: 10),
-                          child: FadeShimmer(
-                            height: 7,
-                            width: 60,
-                            radius: 4,
-                            highlightColor: Color(0xffc0c3d3),
-                            baseColor: Color(0xff999aa1),
-                          ),
-                        )
-                      : Text(
-                          role,
-                          textAlign: TextAlign.center,
-                          style: TextStyle(
-                            color: Theme.of(context).canvasColor,
-                            fontSize: 10,
-                          ),
-                        ),
-                  const SizedBox(height: 4),
-                  imageLink.isEmpty
-                      ? const Padding(
-                          padding: EdgeInsets.symmetric(horizontal: 10),
-                          child: FadeShimmer(
-                            height: 7,
-                            width: 100,
-                            radius: 4,
-                            highlightColor: Color(0xffc0c3d3),
-                            baseColor: Color(0xff999aa1),
-                          ),
-                        )
-                      : Text(
-                          name,
-                          textAlign: TextAlign.center,
-                          style: TextStyle(
-                            color: Theme.of(context).canvasColor,
-                            fontSize: 13,
-                            fontWeight: FontWeight.w900,
-                          ),
-                        ),
-                  const SizedBox(height: 7),
-                  imageLink.isEmpty
-                      ? const Padding(
-                          padding: EdgeInsets.symmetric(horizontal: 10),
-                          child: FadeShimmer(
-                            height: 40,
-                            width: 140,
-                            radius: 4,
-                            highlightColor: Color.fromARGB(255, 241, 101, 201),
-                            baseColor: Pallet.accentColor,
-                          ),
-                        )
-                      : Container(
-                          width: 140,
-                          height: 40,
-                          alignment: Alignment.center,
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(5),
-                            boxShadow: [
-                              BoxShadow(
-                                blurRadius: 7,
-                                spreadRadius: 2,
-                                offset: const Offset(0, 3),
-                                color: Colors.black.withOpacity(0.2),
-                              )
-                            ],
-                            color: Pallet.accentColor,
-                          ),
-                          child: const Text(
-                            "Click to Connect",
-                            style: TextStyle(
-                              fontSize: 12,
-                              fontWeight: FontWeight.bold,
-                              color: Colors.white,
-                            ),
-                          ),
-                        ),
-                  const SizedBox(height: 10),
-                ],
-              ),
-            ),
-          ),
-          Align(
-            alignment: AlignmentDirectional.topCenter,
-            child: imageLink.isEmpty
-                ? const Padding(
-                    padding: EdgeInsets.symmetric(horizontal: 10),
-                    child: FadeShimmer(
-                      height: 200,
-                      width: 180,
-                      radius: 4,
-                      highlightColor: Color(0xffc0c3d3),
-                      baseColor: Color(0xff999aa1),
-                    ),
-                  )
-                : SizedBox(
-                    width: 180,
-                    height: 200,
-                    child: Stack(
-                      children: [
-                        Container(
-                          decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(18),
-                          ),
-                          child: SizedBox(
-                            height: 200,
-                            width: 210,
-                            child: ClipRRect(
-                              borderRadius: BorderRadius.circular(15),
-                              child: CachedNetworkImage(
-                                fit: BoxFit.cover,
-                                imageUrl: imageLink,
-                                progressIndicatorBuilder: (_, __, ___) {
-                                  return const Padding(
-                                    padding:
-                                        EdgeInsets.symmetric(horizontal: 10),
-                                    child: FadeShimmer(
-                                      height: 200,
-                                      width: 180,
-                                      radius: 4,
-                                      highlightColor: Color(0xffc0c3d3),
-                                      baseColor: Color(0xff999aa1),
-                                    ),
-                                  );
-                                },
-                              ),
-                            ),
-                          ),
-                        ),
-                      ],
-                    ),
-                  ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-// class name_card extends StatelessWidget {
-//   const name_card({
-//     Key? key,
-//     required this.imageLink,
-//     required this.role,
-//     required this.name,
-//   }) : super(key: key);
-
-//   final String imageLink;
-//   final String role;
-//   final String name;
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return Stack(
-//       children: [
-//         Container(
-//           decoration: BoxDecoration(
-//             border: Border.all(color: Pallet.accentColor, width: 3),
-//             borderRadius: BorderRadius.circular(18),
-//           ),
-//           child: SizedBox(
-//             height: 250,
-//             width: 200,
-//             child: ClipRRect(
-//               borderRadius: BorderRadius.circular(15),
-//               child: Image(
-//                 image: CachedNetworkImageProvider(imageLink),
-//                 fit: BoxFit.cover,
-//               ),
-//             ),
-//           ),
-//         ),
-//         Positioned.fill(
-//           child: Align(
-//             alignment: Alignment.bottomCenter,
-//             child: Container(
-//               height: 75,
-//               width: double.infinity,
-//               decoration: BoxDecoration(
-//                 color: Colors.black.withOpacity(0.4),
-//                 borderRadius: const BorderRadius.only(
-//                   topLeft: Radius.circular(15.0),
-//                   topRight: Radius.circular(15.0),
-//                   bottomLeft: Radius.circular(15.0),
-//                   bottomRight: Radius.circular(15.0),
-//                 ),
-//               ),
-//               child: Column(
-//                 crossAxisAlignment: CrossAxisAlignment.center,
-//                 children: [
-//                   Padding(
-//                     padding: const EdgeInsets.all(8.0),
-//                     child: Text(
-//                       role,
-//                       textAlign: TextAlign.center,
-//                       style: const TextStyle(
-//                         color: Colors.white,
-//                         fontSize: 12,
-//                       ),
-//                     ),
-//                   ),
-//                   Padding(
-//                     padding: const EdgeInsets.only(right: 8),
-//                     child: Text(
-//                       name,
-//                       textAlign: TextAlign.center,
-//                       style: const TextStyle(
-//                         color: Colors.white,
-//                         fontSize: 15,
-//                         fontWeight: FontWeight.w900,
-//                       ),
-//                     ),
-//                   ),
-//                   Padding(
-//                     padding: const EdgeInsets.only(right: 8.0),
-//                     child: Container(
-//                       height: 0.5,
-//                       width: 140,
-//                       color: Colors.white,
-//                     ),
-//                   ),
-//                   const SizedBox(height: 10),
-//                   const Padding(
-//                     padding: EdgeInsets.only(right: 8.0),
-//                     child: Text(
-//                       "Click to Connect",
-//                       style: TextStyle(
-//                         color: Colors.white,
-//                         fontSize: 10,
-//                       ),
-//                     ),
-//                   ),
-//                 ],
-//               ),
-//             ),
-//           ),
-//         ),
-//       ],
-//     );
-//   }
-// }
